Refactor execution nodes into modular package
