+++ qmtl/pipeline/execution_nodes/timing.py
@@ -1,0 +1,38 @@
+"""Timing control nodes."""
+
+from __future__ import annotations
+
+from datetime import datetime, timezone
+
+from qmtl.sdk.node import CacheView, Node, ProcessingNode
+from qmtl.sdk.timing_controls import TimingController
+
+from ._shared import latest_entry
+
+
+class TimingGateNode(ProcessingNode):
+    """Gate orders based on market timing rules."""
+
+    def __init__(
+        self, order: Node, *, controller: TimingController, name: str | None = None
+    ) -> None:
+        self.order = order
+        self.controller = controller
+        super().__init__(
+            order,
+            compute_fn=self._compute,
+            name=name or f"{order.name}_timing",
+            interval=order.interval,
+            period=1,
+        )
+
+    def _compute(self, view: CacheView) -> dict | None:
+        latest = latest_entry(view, self.order)
+        if latest is None:
+            return None
+        ts, order = latest
+        dt = datetime.fromtimestamp(int(ts), tz=timezone.utc)
+        ok, reason, _ = self.controller.validate_timing(dt)
+        if ok:
+            return order
+        return {"rejected": True, "reason": reason}
